Store surface in its own file
