+++ src/archBuild_sconscript
@@ -13,6 +13,7 @@ sources = [
 	"peakWatcher.cpp",
 	"midiEditorCommands.cpp",
 	"envelopeCommands.cpp",
+	"controlSurface.cpp",
 ]
 
 if env["PLATFORM"] == "win32":
